requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
MarkupSafe==3.0.2
numpy==1.26.4
openai==1.74.0
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.11.3
requests==2.32.3
scikit-learn==1.3.2
scipy==1.11.4
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
python-dotenv==1.0.1

